requirements.txt
requests==2.31.0
gitpython==3.1.31
Pillow==10.1.0
